requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
entrypoints==0.4
Faker==25.2.0
favicon==0.7.0
fonttools==4.51.0
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
htbuilder==0.6.2
html5lib==1.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.2.0
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
webencodings==0.5.1
